backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
pillow==10.2.0
django-storages==1.14.2
boto3==1.34.60
botocore==1.34.60
python-dotenv==1.0.0
gunicorn==21.2.0 